pyproject.toml
[project]
name = "lysai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "chromadb>=1.1.1",
    "email-validator>=2.3.0",
    "fastmcp>=2.12.4",
    "google-genai>=1.43.0",
    "httpx>=0.28.1",
    "langgraph>=0.6.10",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.31.0",
    "reportlab>=4.4.4",
]
